pyproject.toml
[tool.poetry]
name = "CPCReady"
version = "0.0.10"
description = "SDK for programing in basic and Basic compile (ugbasic)"
authors = ["Destroyer <destroyer.dcf@gmail.com>"]
readme = "README.md"
license = "GPL-3.0-or-later"
homepage = "https://cpcready.github.io/doc/"
repository = "https://github.com/CPCReady/sdk.git"
documentation = "https://cpcready.github.io/doc/"
keywords = ["packaging", "cpcready","amstrad","programing"]
classifiers = [
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

packages = [
    { include = "CPCReady" }
]

[virtualenvs]
create = true
in-project = true

[tool.poetry.dependencies]
python = "^3.8"
click = ">=8.1.7,<8.2.0"
rich = ">=13.6.0,<13.7.0"
jinja2 = ">=3.1.2,<3.2.0"
requests = ">=2.31.0,<2.32.0"
packaging = ">=23.2,<24.0"
inquirer = ">=3.1.3,<3.2.0"
pyyaml = "^6.0.1"
cerberus = "^1.3.5"
prompt-toolkit = "^3.0.43"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
cpcr     = "CPCReady.main:main"
cpc      = "CPCReady.main:main"
cpcready = "CPCReady.main:main"
save     = "CPCReady.main:save"
SAVE     = "CPCReady.main:save"
run      = "CPCReady.main:run"
RUN      = "CPCReady.main:run"
screen   = "CPCReady.main:screen"
SCREEN   = "CPCReady.main:screen"
SPRITE   = "CPCReady.main:sprite"
sprite   = "CPCReady.main:sprite"
load     = "CPCReady.main:load"
LOAD     = "CPCReady.main:load"
new      = "CPCReady.main:new"
NEW      = "CPCReady.main:new"
merge    = "CPCReady.main:merge"
MERGE    = "CPCReady.main:merge"
palette  = "CPCReady.main:palette"
PALETTE  = "CPCReady.main:palette"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
